Split new Analytics code into new entrypoint and separate from legacy code.
